flask/requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1  
agno  